requirements.txt
vnstock==3.2.2
pandas>=2.0
numpy>=1.24
pandas-ta>=0.3.14b
scikit-learn>=1.3
xgboost>=2.0
lightgbm>=4.0
vectorbt>=0.26
matplotlib>=3.7
seaborn>=0.12
jupyterlab>=4.0
python-dotenv>=1.0
pyarrow>=12.0
